requirements.txt
dj-database-url==0.4.2
Django==1.11.5
django-bootstrap3==9.0.0
django-cors-headers==2.1.0
djangorestframework==3.6.4
gunicorn==19.7.1
psycopg2==2.7.3.1
pytz==2017.2
whitenoise==3.3.0
